cluster-check-py/requirements.txt
kubernetes
openshift-client